# backend/internal/interface/http/handler/comment.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// CommentHandler handles comment-related HTTP requests
type CommentHandler struct {
	// TODO: Add use case dependencies
}

// NewCommentHandler creates a new CommentHandler
func NewCommentHandler(useCase interface{}) *CommentHandler {
	return &CommentHandler{}
}

// List lists comments for a post
// @Summary List comments
// @Description Get all comments for a specific post
// @Tags comments
// @Accept json
// @Produce json
// @Param postId path int true "Post ID"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /comments/post/{postId} [get]
func (h *CommentHandler) List(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "List comments - TODO"})
}

// Create creates a new comment
// @Summary Create comment
// @Description Create a new comment on a post (Anonymous or Authenticated)
// @Tags comments
// @Accept json
// @Produce json
// @Param postId path int true "Post ID"
// @Param request body map[string]interface{} true "Comment creation request"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Failure 429 {object} map[string]interface{}
// @Router /comments/post/{postId} [post]
func (h *CommentHandler) Create(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Create comment - TODO"})
}

// Update updates a comment
// @Summary Update comment
// @Description Update an existing comment (Owner or Admin)
// @Tags comments
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Param request body map[string]interface{} true "Comment update request"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Failure 429 {object} map[string]interface{}
// @Router /comments/{id} [put]
func (h *CommentHandler) Update(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Update comment - TODO"})
}

// Delete deletes a comment
// @Summary Delete comment
// @Description Delete a comment (Owner or Admin)
// @Tags comments
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Success 200 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Failure 429 {object} map[string]interface{}
// @Router /comments/{id} [delete]
func (h *CommentHandler) Delete(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Delete comment - TODO"})
}

// Like likes a comment
// @Summary Like comment
// @Description Add like to a comment (Authenticated users only)
// @Tags comments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Comment ID"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Failure 429 {object} map[string]interface{}
// @Router /comments/{id}/like [post]
func (h *CommentHandler) Like(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Like comment - TODO"})
}

// Unlike unlikes a comment
// @Summary Unlike comment
// @Description Remove like from a comment (Authenticated users only)
// @Tags comments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Comment ID"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Failure 429 {object} map[string]interface{}
// @Router /comments/{id}/like [delete]
func (h *CommentHandler) Unlike(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Unlike comment - TODO"})
}

// ListReplies lists all replies to a comment
// @Summary List replies
// @Description Get all replies to a specific comment (nested comments)
// @Tags comments
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /comments/{id}/replies [get]
func (h *CommentHandler) ListReplies(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "List replies - TODO"})
}

// CreateReply creates a reply to a comment
// @Summary Create reply
// @Description Create a reply to an existing comment (Anonymous or Authenticated)
// @Tags comments
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Param request body map[string]interface{} true "Reply creation request"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Failure 429 {object} map[string]interface{}
// @Router /comments/{id}/replies [post]
func (h *CommentHandler) CreateReply(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Create reply - TODO"})
}